Fix naming of component, refactor a bit
