+++ frontend/src/grocery_list/GroceryCard.js
@@ -20,6 +20,7 @@ import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 import AssignmentTurnedInIcon from '@material-ui/icons/AssignmentTurnedIn';
 import axios from 'axios';
+
 const backend_url = "http://127.0.0.1:9000/"
 
 export default function GroceryCard(props) {
